requirements.txt
pdfplumber
pandas
openpyxl
python-docx